requirements.txt
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
openai==1.55.3